requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
itsdangerous==1.1.0
Jinja2==2.11.2
llvmlite==0.34.0
MarkupSafe==1.1.1
matplotlib==3.3.2
natsort==7.0.1
numba==0.51.2
numpy==1.19.2
pandas==1.1.2
Pillow==7.2.0
python-dateutil==2.8.1
scikit-learn==0.23.2
scipy==1.5.2
toml==0.10.1
umap-learn==0.4.6
Werkzeug==1.0.1

git+https://github.com/ideasman42/isect_segments-bentley_ottmann.git
